+++ src/ide.c
@@ -327,7 +327,7 @@ static void callnonreadcd(IDE *ide);
 static void callreadcd(IDE *ide);
 static void atapicommand(int ide_board);
 
-int idecallback[3] = {0, 0, 0};
+int64_t idecallback[3] = {0, 0, 0};
 
 int cur_ide[3];
